fix circular import by splitting in model and models
